Move mock data to its own file
