+++ src/services/mockBackend.ts
@@ -1,58 +1,15 @@
 import axios from 'axios'
 import AxiosMockAdapter from 'axios-mock-adapter'
-import Notification, { NotificationState } from '../types/Notification'
-import PatientRecord from '../types/PatientRecord'
+
+import mockData from './mockData'
 
 const moxios = axios.create()
 
-export const axiosMockAdapterInstance = new AxiosMockAdapter(moxios, {
+const axiosMockAdapterInstance = new AxiosMockAdapter(moxios, {
   delayResponse: 0
 })
 
-const notifications: Notification[] = [
-  {
-    id: '123',
-    type: 'Threshold',
-    created: 1605112212,
-    names: 'Bob Smith',
-    patient_id: '0x5a',
-    state: NotificationState.New,
-    golden_id: '0x45',
-    score: 0.5
-  }
-]
-
-const patientRecord: PatientRecord = {
-  auxId: '0x66',
-  type: 'Current',
-  score: 0,
-  uid: '0x98',
-  nationalId: '123456',
-  givenName: 'Bob',
-  familyName: 'Smith',
-  gender: 'Male',
-  dob: new Date(1605522215666),
-  phoneNumber: '085145855',
-  city: 'Welkom',
-  updatedBy: undefined
-}
-
-const goldenRecord: PatientRecord = {
-  auxId: '0x77',
-  type: 'Current',
-  score: 0,
-  uid: '0x9833',
-  nationalId: '123456',
-  givenName: 'Bob',
-  familyName: 'Smith',
-  gender: 'Male',
-  dob: new Date(1605522215666),
-  phoneNumber: '085145855',
-  city: 'Welkom',
-  updatedBy: undefined
-}
-
-const goldenRecords = [{ customGoldenRecord: goldenRecord }]
+const { notifications, patientRecord, goldenRecords } = mockData
 
 axiosMockAdapterInstance
   .onGet('/MatchesForReview')
